fix: dissociations in Car, Model and RentalRate entity session beans

+++ CarRentalManagementSystem-ejb/src/java/ejb/session/stateless/CarEntitySessionBean.java
@@ -10,6 +10,7 @@ import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.Query;
+import util.enumeration.StatusEnum;
 import util.exception.CarNotFoundException;
 import util.exception.ModelNotFoundException;
 
@@ -68,9 +69,15 @@ public class CarEntitySessionBean implements CarEntitySessionBeanRemote, CarEnti
     @Override
     public void deleteCar(Long carId) throws CarNotFoundException {
         CarEntity carEntityToRemove = retrieveCarEntityByCarId(carId);
-        em.remove(carEntityToRemove);
+        carEntityToRemove.getModelEntity().getCarEntities().remove(carEntityToRemove);
+        
+        if (carEntityToRemove.getCarStatus() == StatusEnum.USED) {
+            carEntityToRemove.setCarStatus(StatusEnum.DISABLED);
+        } else {
+            em.remove(carEntityToRemove);
+        }
+
         em.flush();
     }
-
     
 }
