+++ CarRentalManagementSystem-ejb/src/java/ejb/session/stateless/RentalRateEntitySessionBean.java
@@ -10,6 +10,7 @@ import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.Query;
+import util.enumeration.StatusEnum;
 import util.exception.CategoryNotFoundException;
 import util.exception.RentalRateNotFoundException;
 
@@ -71,7 +72,14 @@ public class RentalRateEntitySessionBean implements RentalRateEntitySessionBeanR
     @Override
     public void deleteRentalRate(Long rentalRateId) throws RentalRateNotFoundException {
         RentalRateEntity rentalRateEntityToRemove = retrieveRentalRateEntityByRentalRateId(rentalRateId);
-        em.remove(rentalRateEntityToRemove);
+        rentalRateEntityToRemove.getCategoryEntity().getRentalRateEntities().remove(rentalRateEntityToRemove);
+        
+        if (rentalRateEntityToRemove.getRentalRateStatus() == StatusEnum.USED) {
+            rentalRateEntityToRemove.setRentalRateStatus(StatusEnum.DISABLED);
+        } else {
+            em.remove(rentalRateEntityToRemove);
+        }
+        
         em.flush();
     }
 }
